Cat Rescue Mock Homepage
This was an assignment to work with a color palette of primary, secondary and accent colors based off emotions of happiness, professionalism and joy. Also the assignment was to include google font for typography and build a simple homepage using a stock cat photo. Elements to include were a header, image, body, button and footer.